Kmgnw (#113)
* ✨ 인풋창 클릭 시 화면 확대 방지 코드 추가

* ✨ 로그인 검증 전역 함수 추가

* ✨ 로그인 조건부 렌더링 로직 추가

* ✨ 마이페이지 커뮤니티 셀 수정

* ✨ 채팅 완성

* ✨ 인풋 플레이스홀더 변경경ㅐ

* 🐛 피드백 수정

---------

+++ frontend/src/entities/AddSchedule/api/api.js
@@ -1,4 +1,5 @@
 import { baseUrl } from "../../../shared/config/baseurl";
+import { isLogin } from "../../../shared/function/isLogin";
 
 export const postSchedule = async (newSchedule) => {
     return fetch(`${baseUrl}/api/schedules`, {
@@ -8,40 +9,52 @@ export const postSchedule = async (newSchedule) => {
             'Authorization': `Bearer ${window.sessionStorage.getItem('token')}`
         },
         body: JSON.stringify(newSchedule),
-      })
-      .then((response) => response.json())
-      .then((data) => {
-        console.log('일정등록완')
-        console.log(data.result.schedule_id)
-      })
-      .catch((error) => {
-          console.error('Error:', error);
-      });
+    })
+        .then((response) => response.json())
+        .then((data) => {
+            console.log('일정등록완')
+            console.log(data.result.schedule_id)
+        })
+        .catch((error) => {
+            console.error('Error:', error);
+        });
 };
 
-export function deleteSchedule(schedule_id, ) {
-    const token = window.sessionStorage.getItem('token');
-    const url = `${baseUrl}/api/schedules/${schedule_id}`;
+export function deleteSchedule(schedule_id,setAugust,crntClickedDay, crntClickedIndexOfSchedules) {
 
-    return fetch(url, {
-        method: 'DELETE',
-        headers: {
-            'Authorization': `Bearer ${token}`, // 필요한 경우 인증 헤더 추가
-            'Content-Type': 'application/json'  // 필요에 따라 헤더 추가
-        }
-    })
-    .then(response => {
-        if (!response.ok) {
-            return response.json().then(error => { throw new Error(error.detail); });
-        }
-        return response.json();
-    })
-    .then(data => {
-        console.log('delete', data);
-        return data;
-    })
-    .catch(error => {
-        console.error('Error:', error.message);
-        throw error;
-    });
+    if (isLogin()) {
+
+        const token = window.sessionStorage.getItem('token');
+        const url = `${baseUrl}/api/schedules/${schedule_id}`;
+
+        return fetch(url, {
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${token}`, // 필요한 경우 인증 헤더 추가
+                'Content-Type': 'application/json'  // 필요에 따라 헤더 추가
+            }
+        })
+            .then(response => {
+                if (!response.ok) {
+                    return response.json().then(error => { throw new Error(error.detail); });
+                }
+                return response.json();
+            })
+            .then(data => {
+                console.log('delete', data);
+                return data;
+            })
+            .catch(error => {
+                console.error('Error:', error.message);
+                throw error;
+            });
+    }else{
+        setAugust((prev) => {
+            const list = JSON.parse(JSON.stringify(prev));
+            list[crntClickedDay].splice(crntClickedIndexOfSchedules, 1);
+            return list;
+        });
+    }
+
+
 }
